Reimplement the RSA ASN1 types and move file to crypto/rsa.
This makes use of the BIGNUM and LONG ASN1 types
and the new/free overrides.

Reimplement NETSCAPE_PKEY. The rest of the "NET"
form will be trickier due to broken encoding issues
with IIS and because the current code manually handles
the encoding in a convoluted way.

Zap all the old ASN1 error codes and rebuild. This
has deleted lots of old error codes because the
new ASN1 doesn't use two function codes per structure.

+++ crypto/rsa/rsa.h
@@ -113,7 +113,7 @@ struct rsa_st
 	/* The first parameter is used to pickup errors where
 	 * this is passed instead of aEVP_PKEY, it is set to 0 */
 	int pad;
-	int version;
+	long version;
 #if 0
 	RSA_METHOD *meth;
 #else
